Refactor Game Section to separate component
