+++ src/lib/index.ts
@@ -1,5 +1,5 @@
 export { default as PineappleLayoutBase } from "./components/pineapple/PineappleBaseLayout.svelte";
-export { default as SeaweedTemplate } from "./template/SeaweedTemplate.svelte";
+export { default as SeaweedTemplate } from "./template/Seaweed/SeaweedTemplate.svelte";
 export { default as LazyAsset } from "./components/LazyAsset.svelte";
 export { default as Card } from "./components/Card.svelte";
